Move `ImageGroupFixtures` to the proper package
